requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bootstrap
Flask-Login
Flask-Mail
Flask-Migrate
Flask-Principal
Flask-Moment
Flask-PageDown
Flask-WTF
python-dateutil
python-dotenv
WTForms
email_validator
flask-change-password
Werkzeug==2.2.2
fpdf
Flask-Babel
firebase-admin